# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /royalties:
    get:
      operationId: listRoyalties
      tags: [royalties]
      summary: List royalty records
      security:
        - bearerAuth: []
      parameters:
        - name: artistId
          in: query
          schema:
            type: integer
        - name: releaseId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of royalties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Royalty"
    post:
      operationId: createRoyalty
      tags: [royalties]
      summary: Create a royalty record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoyaltyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Royalty"
  /royalties/{id}:
    get:
      operationId: getRoyalty
      tags: [royalties]
      summary: Get royalty record by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Royalty record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Royalty"
    put:
      operationId: updateRoyalty
      tags: [royalties]
      summary: Update royalty record (e.g. mark as paid)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoyaltyBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Royalty"
    delete:
      operationId: deleteRoyalty
      tags: [royalties]
      summary: Delete a royalty record
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CONTENT POSTS ──────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Royalty:
      type: object
      properties:
        id:
          type: integer
        artistId:
          type: integer
        releaseId:
          type: integer
          nullable: true
        periodStart:
          type: string
        periodEnd:
          type: string
        streamCount:
          type: integer
        grossCents:
          type: integer
        netCents:
          type: integer
        splitPct:
          type: integer
        status:
          type: string
          enum: [pending, processing, paid, disputed]
        paidAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, artistId, periodStart, periodEnd, streamCount, grossCents, netCents, splitPct, status, createdAt]
    RoyaltyBody:
      type: object
      properties:
        artistId:
          type: integer
        releaseId:
          type: integer
        periodStart:
          type: string
        periodEnd:
          type: string
        streamCount:
          type: integer
        grossCents:
          type: integer
        netCents:
          type: integer
        splitPct:
          type: integer
        status:
          type: string
          enum: [pending, processing, paid, disputed]
        notes:
          type: string
      required: [artistId, periodStart, periodEnd, streamCount, grossCents, netCents, splitPct]
...